sql/expression/function: move makedate to dedicated files
